fix type of Relations tables
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@87068 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: df6686329fc892e477737ed5092860e83a46e618

+++ Rec/Calo/CaloMoniDst/src/CaloPhotonChecker.cpp
@@ -327,9 +327,9 @@ StatusCode CaloPhotonChecker::finalize()
 // ============================================================================
 StatusCode CaloPhotonChecker::execute()
 { 
-  typedef IRelationWeighted<LHCb::CaloCluster,LHCb::MCParticle,double>  MCTable;
+  typedef IRelationWeighted<LHCb::CaloCluster,LHCb::MCParticle,float>  MCTable;
   typedef MCTable::Range                                               MCRange;
-  typedef IRelationWeighted<LHCb::CaloCluster,LHCb::Track,double>       Table;
+  typedef IRelationWeighted<LHCb::CaloCluster,LHCb::Track,float>       Table;
   typedef Table::Range                                                 Range;
 
   typedef Gaudi::Math::Line<Gaudi::XYZPoint,Gaudi::XYZVector> Line;
